feat: enable child bus reconnects
This change allows a parent (hosting) page to listen for reconnections
from a child EventBus instance beyond the initial connection. This
situation can happen when (e.g. the child iframe connects to the parent
but then needs to navigate away; and eventually the iframe settles on a
page that needs to re-establish a connection to the parent bus).